requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-api-python-client
streamlit-quill
